requirements.txt
python-dotenv==1.0.1
streamlit==1.38.0
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.2
chromadb==0.5.5
pypdf==5.0.0
markdownify==0.12.1
google-generativeai==0.7.2
